+++ src/pages/CreatePostPage/CreatePostSuccessModal/index.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import styled from '@emotion/styled';
-import LinkButton from '@components/NavBar/LinkButton';
+import LinkButton from '@components/LinkButton';
 
 interface CreatePostSuccessModalProps {
   postId: string | null;
